# Rework GroupByHash to for faster performance and support grouping by nulls