# Which issue does this PR close? 

Closes https://github.com/apache/arrow-datafusion/issues/790 by implementing a new design for group by hash. Built on https://github.com/apache/arrow-datafusion/pull/812 so it may be easier to review that one first

This PR is an amazing collaborative effort and includes ideas from @Dandandan  @jhorstmann @rdettai @jorgecarleitao and likely others I forgot.

 # Rationale for this change
1. Regain performance lost when we added support for GROUP BY NULL; See https://github.com/apache/arrow-datafusion/issues/790 for more details

# What changes are included in this PR?
1. Use a hash to to create the appropriate grouping, use indexes rather than hash keys many time

# Potential Follow On Work
* There are some great ideas from @sundy-li  and @Dandandan on https://github.com/apache/arrow-datafusion/issues/846#issuecomment-895763475 for additional ways to improve grouping performance

# Performance Summary
* db-query: @Dandandan 's measurements show improvement except for one query 
* [DataFusion Synthetic Aggregate Benchmark](https://github.com/alamb/datafusion_aggregate_bench) (specially created for this work): %15 - 27% improvement
* Aggregate microbenches show no changes within my margin of error

## db-query
The numbers @Dandandan measureed are at  https://github.com/apache/arrow-datafusion/pull/808#issuecomment-895425959

## [DataFusion Synthetic Aggregate Benchmark](https://github.com/alamb/datafusion_aggregate_bench) (newly created for this work): 

Full results, https://github.com/alamb/datafusion_aggregate_bench/blob/main/RESULTS.md . Summary:

|test                                                                                    | master  | [arrow-datafusion #808](https://github.com/apache/arrow-datafusion/pull/808)| `gby_null` / `master` (less than 1 is better) |
| ----------- | ----------- |------------ | ------- |
| `100 Groups; 100M rows, int64_keys(10% nulls), f64 values(1% nulls)`                   | 22.40s  | 16.27s | .73  |
| `100 Groups; 100M rows, utf8_keys(10% nulls),  f64 values(1% nulls)`                   | 29.46s  | 22.73s | .77  |
| `100 Groups; 100M rows, dictionary(utf8, int32) keys(10% nulls), f64 values(1% nulls)` | 31.54s  | 26.96s | .85  |

## aggregate micro benchmarks
Tested via
```
cargo bench --bench aggregate_query_sql -- --save-baseline <test name>
```

Results
```
group                                                gby_null_new1                           master
-----                                                -------------                           ------
aggregate_query_group_by                             1.00      2.9±0.13ms        ? ?/sec     1.10      3.2±0.35ms        ? ?/sec
aggregate_query_group_by_u64 15 12                   1.00      3.1±0.16ms        ? ?/sec     1.00      3.1±0.08ms        ? ?/sec
aggregate_query_group_by_with_filter                 1.12      2.2±0.12ms        ? ?/sec     1.00  1969.5±41.36µs        ? ?/sec
aggregate_query_group_by_with_filter_u64 15 12       1.09      2.2±0.08ms        ? ?/sec     1.00      2.1±0.08ms        ? ?/sec
aggregate_query_no_group_by 15 12                    1.00  1195.8±70.21µs        ? ?/sec     1.02  1223.4±108.78µs        ? ?/sec
aggregate_query_no_group_by_count_distinct_narrow    1.00      5.5±0.23ms        ? ?/sec     1.09      6.0±0.90ms        ? ?/sec
aggregate_query_no_group_by_count_distinct_wide      1.09      8.0±0.59ms        ? ?/sec     1.00      7.4±0.30ms        ? ?/sec
aggregate_query_no_group_by_min_max_f64              1.08  1177.6±117.36µs        ? ?/sec    1.00  1092.1±29.93µs        ? ?/sec
```


## Performance Source
This approach avoids the following operations which should improve its speed:
1. Avoids copying GroupValues into a Vec to hash, saving both time and space
2. Avoids several hash table lookups (used indexes into `group_values` instead
3. Uses vectorized hashing

# Are there any user-facing changes?
Faster performance


# Notes:
I tried to keep the same names and structure of the existing hash algorithm (as I found that easy to follow -- nice work @Dandandan  and @andygrove ) and I think that will make this easier to review

# Items completed
- [x] Basic debugging
- [x] ~Add ascii art diagrams~ (the design got simpler using `RawTable` API)
- [x] Measure performance
- [x] Handle hash collisions - https://github.com/apache/arrow-datafusion/issues/843
- [x] https://github.com/apache/arrow-datafusion/pull/812
- [x] Implement fast comparison with `ScalarValue` and `ArrayRef` https://github.com/apache/arrow-datafusion/pull/844