@alamb 2021-08-09
I am basically done with this PR. All that remains in my mind is to run some benchmarks and I'll mark it as ready for review