@Dandandan 2021-08-09
On the db-benchmark aggregation queries:

PR:
```
q1 took 33 ms
q2 took 377 ms
q3 took 986 ms
q4 took 47 ms
q5 took 973 ms
q7 took 932 ms
q10 took 4040 ms
```

Master:
```
q1 took 37 ms
q2 took 325 ms
q3 took 1431 ms
q4 took 56 ms
q5 took 1287 ms
q7 took 1304 ms
q10 took 9380 ms
```

It looks like it's a small perf hit on q2,  but I think the other 4 queries do greatly compensate for this.