@alamb 2021-08-10
Thanks @Dandandan  and @jorgecarleitao  -- I plan to merge https://github.com/apache/arrow-datafusion/pull/812 in first and leave this one open for another few days in case anyone else wants to comment.