@alamb 2021-08-11
Rebased now that https://github.com/apache/arrow-datafusion/pull/812 has been merged