@Dandandan 2021-08-11
A TCP-H query that also get's quite a bit faster is q13, on Parquet 37.8s -> 29.5s